+++ cluster/cluster.go
@@ -24,6 +24,7 @@ func Register(rcvr interface{}) {
 	cServer.server.Register(rcvr)
 }
 
+/*
 var cClient map[string]*rpc.Client
 
 func Connect(serverName string, addr string) {
@@ -52,3 +53,4 @@ func Send(serverName string, v ...interface{}) {
 	}
 	client.Send(v)
 }
+*/