+++ gate/gateserver.go
@@ -4,21 +4,22 @@ import (
 	"eegos/log"
 
 	"net"
-	"strings"
+	"time"
 )
 
 type Handler interface {
-	Connect(s *Session)
-	Message(data *Data)
-	Close(s *Session)
+	Connect(uint16, chan *Data)
+	Message(uint16, uint16, []byte)
+	HeartBeat(uint16, uint16)
+	Close(uint16)
 }
 
 type TcpServer struct {
 	isOpen bool
-	handle *Handler
+	handle Handler
 }
 
-func NewTcpServer(handle *Handler) *TcpServer {
+func NewTcpServer(handle Handler) *TcpServer {
 	return &TcpServer{isOpen: true, handle: handle}
 }
 
@@ -50,9 +51,43 @@ func (this *TcpServer) Open(addr string) {
 	}()
 }
 
-func (this *TcpServer) NewSession(conn net.TCPConn) {
+func (this *TcpServer) NewSession(conn *net.TCPConn) {
 	log.Debug("new connection from ", conn.RemoteAddr())
 	session := CreateSession(conn)
 	//handle new connection
-	this.handle.Connect(session)
+	this.handle.Connect(session.fd, session.outData)
+	session.Start()
+
+	go this.processData(session)
+}
+
+func (this *TcpServer) processData(s *Session) {
+	for {
+		select {
+		case data, ok := <-s.inData:
+			if !ok {
+				continue
+			}
+			switch data.dType {
+			case HEARTBEAT:
+				go s.Write(data.Head, HEARTBEAT_RET, []byte{})
+				go this.handle.HeartBeat(s.fd, data.Head)
+			case HEARTBEAT_RET:
+				go this.handle.HeartBeat(s.fd, data.Head)
+			default:
+				go this.handle.Message(s.fd, data.Head, data.Body)
+			}
+		case data, ok := <-s.outData:
+			if !ok {
+				continue
+			}
+			go s.Write(data.Head, DATA, data.Body)
+
+		case <-s.cClose:
+			log.Debug("session close")
+			this.handle.Close(s.fd)
+			s.Release()
+			break
+		}
+	}
 }