+++ rpc/server.go
@@ -1,11 +1,12 @@
 package rpc
 
 import (
-	"log"
-	"net"
+	"eegos/gate"
+	"eegos/log"
+
+	"encoding/json"
 	"reflect"
 	"strings"
-	//"time"
 )
 
 type methodType struct {
@@ -22,106 +23,101 @@ type Service struct {
 
 type Server struct {
 	serviceMap map[string]*Service
-	//sessions   map[string]*Session
-	//session    *Session
-	closed bool
+	tcpGate    *gate.TcpServer
+	outDatas   map[uint16]chan *gate.Data
 }
 
 func NewServer() *Server {
 	services := make(map[string]*Service)
-	//sessions := make(map[string]*Session)
-	return &Server{serviceMap: services, closed: false}
+	newServer := Server{serviceMap: services}
+	newServer.tcpGate = gate.NewTcpServer(&newServer)
+	newServer.outDatas = make(map[uint16]chan *gate.Data)
+
+	return &newServer
+}
+
+func (this *Server) Connect(fd uint16, outData chan *gate.Data) {
+	log.Debug("rpc server new connection", fd)
+	this.outDatas[fd] = outData
+}
+
+func (this *Server) Message(fd uint16, sessionID uint16, body []byte) {
+	outc, ok := this.outDatas[fd]
+	if !ok {
+		log.Error("fd not found", fd, sessionID)
+		return
+	}
+
+	args := []interface{}{}
+	err := json.Unmarshal(body, &args)
+	if err != nil {
+		log.Error(err)
+		return
+	}
+
+	if len(args) == 0 {
+		log.Error("no args")
+		return
+	}
+
+	info := args[0].(string)
+
+	dot := strings.LastIndex(info, ".")
+	if dot < 0 {
+		log.Error("cannot find dot")
+		return
+	}
+	serviceName := info[:dot]
+	methodName := info[dot+1:]
+
+	sInfo := this.serviceMap[serviceName]
+	if sInfo == nil {
+		log.Error("cannot find service")
+		return
+	}
+	mInfo := sInfo.methods[methodName]
+	if mInfo == nil {
+		log.Error("cannot find method", methodName, len(methodName))
+		return
+	}
+
+	callArgs := make([]reflect.Value, len(mInfo.args)+1)
+	callArgs[0] = sInfo.rcvr
+
+	for i := 0; i < len(mInfo.args); i++ {
+		/*
+			if reflect.TypeOf(args[i+1]).Kind() != mInfo.args[i].Kind() {
+				callArgs[i+1] = reflect.ValueOf(args[i+1]).Convert(mInfo.args[i])
+				log.Println("arg type convert")
+			} else {
+				callArgs[i+1] = reflect.ValueOf(args[i+1])
+			}
+		*/
+		callArgs[i+1] = reflect.ValueOf(args[i+1]).Convert(mInfo.args[i])
+	}
+	ret := this.RunFunc(mInfo, callArgs)
+	retBody, err := json.Marshal(ret)
+	if err != nil {
+		log.Error(err)
+		return
+	}
+	retPkg := &gate.Data{Head: sessionID, Body: retBody}
+	outc <- retPkg
+}
+
+func (this *Server) HeartBeat(fd uint16, sessionID uint16) {
+	log.Debug("HeartBeat", fd, sessionID)
+}
+
+func (this *Server) Close(fd uint16) {
+	log.Debug("rpc server close", fd)
 }
 
 func (this *Server) Open(addr string) {
-	tcpAddr, err := net.ResolveTCPAddr("tcp4", addr)
-	if err != nil {
-		log.Println("cluster.Open: net.ResolveTCPAddr: ", err.Error())
-		return
-	}
-	lis, err := net.ListenTCP("tcp", tcpAddr)
-
-	log.Println("cluster.Open: listening")
-	if err != nil {
-		log.Println("cluster.Open: net.ListenTCP: ", err.Error())
-		return
-	}
-
-	go func() {
-		for !this.closed {
-			conn, err := lis.Accept()
-			if err != nil {
-				log.Println("cluster.Open: lis.Accept: ", err.Error())
-				return
-			}
-			//conn.SetKeepAlive(true)
-			//conn.SetKeepAlivePeriod(5 * time.Second)
-			go this.NewSession(conn)
-		}
-	}()
+	this.tcpGate.Open(addr)
 }
 
-func (this *Server) NewSession(conn net.Conn) {
-	log.Println("new connection")
-	session := CreateSession(conn)
-	//this.session = session
-	//this.sessions[conn.RemoteAddr().String()] = session
-	for {
-		select {
-		case data, ok := <-session.newData:
-			if !ok {
-				continue
-			}
-			args := data.body
-			session_id := data.head
-			if len(args) == 0 {
-				continue
-			}
-
-			info := args[0].(string)
-
-			dot := strings.LastIndex(info, ".")
-			if dot < 0 {
-				log.Println("cannot find dot")
-				continue
-			}
-			serviceName := info[:dot]
-			methodName := info[dot+1:]
-
-			sInfo := this.serviceMap[serviceName]
-			if sInfo == nil {
-				log.Println("cannot find service")
-				continue
-			}
-			mInfo := sInfo.methods[methodName]
-			if mInfo == nil {
-				log.Println("cannot find method", methodName, len(methodName))
-				continue
-			}
-
-			callArgs := make([]reflect.Value, len(mInfo.args)+1)
-			callArgs[0] = sInfo.rcvr
-
-			for i := 0; i < len(mInfo.args); i++ {
-				/*
-					if reflect.TypeOf(args[i+1]).Kind() != mInfo.args[i].Kind() {
-						callArgs[i+1] = reflect.ValueOf(args[i+1]).Convert(mInfo.args[i])
-						log.Println("arg type convert")
-					} else {
-						callArgs[i+1] = reflect.ValueOf(args[i+1])
-					}
-				*/
-				callArgs[i+1] = reflect.ValueOf(args[i+1]).Convert(mInfo.args[i])
-			}
-			go this.RunFunc(mInfo, callArgs, session, session_id)
-		case <-session.cClose:
-			log.Println("session close")
-			break
-		}
-	}
-}
-
-func (this *Server) RunFunc(m *methodType, args []reflect.Value, session *Session, session_id uint16) {
+func (this *Server) RunFunc(m *methodType, args []reflect.Value) []interface{} {
 	callRet := m.method.Func.Call(args)
 
 	retLen := len(callRet)
@@ -129,7 +125,9 @@ func (this *Server) RunFunc(m *methodType, args []reflect.Value, session *Sessio
 	for i := 0; i < retLen; i++ {
 		reply[i] = callRet[i].Interface()
 	}
-	session.HandleWrite(session_id, reply)
+
+	return reply
+	//session.HandleWrite(sessionID, reply)
 }
 
 func (this *Server) Register(rcvr interface{}) {
@@ -160,8 +158,8 @@ func (this *Server) Register(rcvr interface{}) {
 			}
 		}
 		s.methods[mname] = &methodInfo
-		log.Println("registered method: ", mname, len(methodInfo.args))
+		log.Debug("registered method: ", mname, len(methodInfo.args))
 	}
 	this.serviceMap[s.name] = s
-	log.Println("registered Service: ", s.name)
+	log.Debug("registered Service: ", s.name)
 }