+++ CHANGELOG.md
@@ -5,11 +5,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.0.2] - 2019-12-31
+### Changed
+- keep session simple
+- keep session inside package
+- process heartbeat in listening server
+### Added
+- gate handle 
+- new package util
+- new rpc use gate
+- outData channel for write buff
+### Deprecated
+- old rpc
+
 ## [0.0.2] - 2019-12-29
 ### Changed
-- use go mod
 - make session as gate session
-- new package: gate, log
+### Added
+- use go mod
+- new package gate, log
 
 ## [0.0.1] - 2018-06-17
 ### Changed